Clean up types to be more portable

+++ mcp47feb.h
@@ -24,12 +24,12 @@ void 	            MCP47FEB_Init(MCP47FEB_TypeDef *dac, uint8_t devAddr, I2C_Hand
 HAL_StatusTypeDef   MCP47FEB_IsReady(MCP47FEB_TypeDef *dac);
 void  	            MCP47FEB_UnlockSALCK(MCP47FEB_TypeDef *dac);
 void  	            MCP47FEB_LockSALCK(MCP47FEB_TypeDef *dac, uint8_t addr);
-uint8_t             MCP47FEB_GetPowerDown(MCP47FEB_TypeDef *dac, int channel);
-void 	            MCP47FEB_SetPowerDown(MCP47FEB_TypeDef *dac, int val0, int val1);
-uint8_t             MCP47FEB_GetPowerDownEp(MCP47FEB_TypeDef *dac, int channel);
-uint8_t             MCP47FEB_GetGain(MCP47FEB_TypeDef *dac, int channel);
-void 	            MCP47FEB_SetGain(MCP47FEB_TypeDef *dac, int val0, int val1);
-uint8_t             MCP47FEB_GetGainEp(MCP47FEB_TypeDef *dac, int channel);
+uint8_t             MCP47FEB_GetPowerDown(MCP47FEB_TypeDef *dac, uint8_t channel);
+void 	            MCP47FEB_SetPowerDown(MCP47FEB_TypeDef *dac, uint8_t val0, uint8_t val1);
+uint8_t             MCP47FEB_GetPowerDownEp(MCP47FEB_TypeDef *dac, uint8_t channel);
+uint8_t             MCP47FEB_GetGain(MCP47FEB_TypeDef *dac, uint8_t channel);
+void 	            MCP47FEB_SetGain(MCP47FEB_TypeDef *dac, uint8_t val0, uint8_t val1);
+uint8_t             MCP47FEB_GetGainEp(MCP47FEB_TypeDef *dac, uint8_t channel);
 uint8_t             MCP47FEB_GetVref(MCP47FEB_TypeDef *dac, uint8_t channel);
 void 	            MCP47FEB_SetVref(MCP47FEB_TypeDef *dac, uint8_t val0, uint8_t val1);
 uint8_t             MCP47FEB_GetVrefEp(MCP47FEB_TypeDef *dac, uint8_t channel);
